Moved the PrintEvents visitor into Viewer::pollEvents() to facilitate test across Window implementations. Cleaned up the XcbWindow constructor in prep for implementing the Vulkan setup.
